Add enhanced tooltips with relative and absolute values.
The mouse-over tooltips for most graphs now provide values relative to
the previous day, helping to visually see the trends when scanning
through data.

The "% of Cases In Each Batch Of Test Results" graph also shows both the
absolute counts and the percentage in the tooltip, providing more
meaning for those who are drilling in.