Make it easier to scroll the page on mobile devices.
It was pretty annoying scrolling through the pages on iPads/iPhones and
other mobile devices, since the graphs took up the entire width of the
page and would trigger graph zoom behavior on any touch events. We now
have more spacing around these to give a consistent vertical column on
either side of the viewport for scrolling.

It's not perfect, but it's better than what I had.