Add a new data feed scraper for the Butte County dashboard.
This makes an attempt to pull in data from the Butte County dashboard
by deserializing the Infogram chart data, locating the proper entities,
verifying value, and extracting data.

This can definitely fail if they change the structure, but there's a
whole lot of safeguards in place to make sure we don't end up with
corrupted data. At worst, we should simply not write out new data,
requiring adjustments to the script to continue after that point.

Initially, this gets outputted to JSON, and all history is preserved. A
CSV feed is then generated based on the JSON data.