Add a small pillow behind the event information text in the timelines.
The event information text could be hard to see when overlaying a bar in
the graph. While this usually works itself out as the number of cases
grows, it's not very helpful in the meantime. To make this just a tad
easier to see, there's now a soft, semi-transparent stroke around the
text to help cut into the bar. It's not perfect, but helps a little with
readability.