Don't trust the Butte County dashboard date stamps.
A case was hit today where Butte County updated data but not the date,
and this resulted in some data for yesterday getting overridden.
Fortunately, we have backups, but this is annoying and shouldn't be
tolerated.

We now validate the data coming from the page. If it isn't today, skip
any further processing of the page. This should also result in faster
scanning, in the end.