Add an optional panel for selecting date ranges.
This removes the zoom interaction capabilities of the graphs, which were
getting in the way of scrolling and interacting with the pages on mobile
devices, and weren't obvious.

Instead, we now have a docked option bar on the bottom of the page for
selecting date ranges. These have some built-in defaults, and an option
for specifying an explicit range of dates.

Once any option is selected, all timeline graphs on the page will update
accordingly.