Add HTTP caching for feeds, to reduce impact on servers.
We now keep a HTTP cache file that stores ETags for feeds, preventing us
from having to fetch data we've already fetched.

This mainly applies to ones hosted on places like GitHub. It sadly does
not apply to infogram.com, home of the Butte County dashboard, as this
has nonsensical (completely busted and useless) HTTP caching.