Use a better date representation for the counters.
The counters had hard-coded "since yesterday" and "since previous day"
to account for different update schedules in the data feeds, but these
weren't really that useful, or meaningful on weekends.

The dashboard now computes a proper value for each of these labels.
It will return "today", "yesterday", or a day of the week, depending on
what's appropriate. This should help ensure that it's clear what day a
number is based off.