Add an indicator to unchanged cases and grey them out.
The "0 today" or "0 yesterday" indicators under a counter were a little
confusing. It could make it seem like the value was 0 yesterday, or
whatever. To help make it clear, we now have the up arrow indicator
before it, showing it's a relative value. We also grey it out, so that
the eyes can focus on what matters on the page, reducing noise and
making it more clear that there's nothing to look at there.