Move to a new webpack-based build and packaging setup.
Since we're launching soon, it's about time to move to a proper
versioned, minifed setup for media assets.

The tool chosen here is webpack, which I have mixed feelings about, but
in the end it works. The source code for the site now lives in src/, and
its built files are put into htdocs/.

The JavaScript is run through Babel, providing wider browser
compatibility, and CSS is run through Less. The CSS will be moved fully
over to Less in an upcoming change.