Parse county resource data from the State dashboard.
This provides a new state-resources.json/csv dataset that contains the
resources found on the covid19.ca.gov dashboard for the county. This
includes beds, face shields, gloves, gowns, procedure masks, N95
respirators, and the percentage of ICU beds and ventilators available
(though it's unclear from their data whether this is a daily value or a
difference between 2 weeks).