Fix loading of state resource data and improve performance.
The data dictionary for the state resources Tableau dashboard shifted
today, and it's now clear that some of the leading values will disappear
if they're not unique. This means we can't index things from positions
relative to the beginning of the array, but instead need to index
relative to the end (as these are less likely to be duplicated -- though
this is still possible...).