Add a spinner to show while the page is loading.
This helps provide some feedback that the page isn't just broken, which
can happen on a slow connection. We now have a version of the logo that
acts as a spinner, which will be centered horizontally and vertically on
the page while the rest of the content is loading, removed once we're
ready to display the dashboard.