Avoid cutting off the current day's bar in the graphs.
The current day's bar was getting cut off in the standard and day range
views, and the label was getting cut off in the standard view. This
seems to be a byproduct of the zooming support.

To work around this, there's now some extra padding on the right of the
graphs, and the zoom domain has 8 hours added to it, leaving a bit more
room to work with.