Rewrite and re-activate the state resources datasets.
I've figured out (I think) how to reliably get the data out of Tableau
for the resources without having to guess at indexes. For future people
reading this, the aliasIndices (when available and not negative numbers)
seem to be indices into the dataValues of the type matching the column's
dataType field.

We now look up by the sheet and field captions, sanity-checking the
types in the process, and piece the data together to get the state
resources for the day.