Add an automated dataset for the per-hospital case counts.
This is the last big piece of the puzzle for automation of the
dashboard. We now extract the per-hospital case counts from the state
dashboard, giving us a total daily count (trailing one day behind) for
Enloe Hospital, Oroville Hospital, and Orchard Hospital (which doesn't
seem to have reported a case to the state yet, unless I've missed it).