Add a note that timeline events are not claiming to be sources of spread.
This has come up in two separate discussions, one defending the church
gathering, one defending the George Floyd protests. The intent of these
events is to shape the timeline, highlight when cases occurred relative
to dates we can contextualize in our heads, but as we do not have
information on the source of spreads, we're not making any claims that
these are relevant.

The note also points out that regular activities, like going to bars,
having parties, or going to farmers markets, can cause spreads as well.