Update to use the new California COVID-19 hospitals datasets.
A few days ago, California replaced the old hospitals dataset with a
new, expanded one. This contains not only total counts on a given day,
but deltas from previous days that cover new patients admitted for
COVID-19 symptoms and patients converted to COVID-19 status. The only
data that's missing from the old dataset was the death count, though we
don't use that for our purposes.