Redo the header to link to the FAQ and provide info on when things update.
There's been some questions about when information is updated. The
header of the page now goes into this, so people have an idea of what to
expect.

It also links to a FAQ hosted on Notion. This covers a lot of questions
that have been asked so far by followers of the dashboard.