Fix a regression in pulling in hospital counts due to state data changes.
The state reworked their dashboard for county hospital information. They
moved it to a new URL, changed some internal keys for accessing the map
information, and changed data types from `integer` to `real`. All of
this broke any lookups we'd want to do.

This updates the parser to pull that information in.