Switch the SNF graph to a bar graph.
This graph was previously a stacked area graph that grouped the staff
and patient values together, adding them up. It followed two other
stacked area graphs that did not combine values. I felt that was
misleading, since there was no clear indication that the graphs were
doing different things with the same format. It was also inconsistent
with other case graphs.

This is now a bar graph, which should be a lot easier to mentally
process after seeing the other case bar graphs.