Embed the bundled spinner.svg into the CSS at compile time.
Prior to using LessCSS, I wanted to embed the spinner SVG in order to
prevent having to fetch the image before the spinner could be shown.
This was done by manually constructing a data URI.

Now that I'm using LessCSS, it makes sense to just embed at compile time
from the source image, making it easier to maintain the image.