Add separate counters and graphs for hospitalized residents.
One of the most commonly-asked questions is why the hospitalization
count differs between bc19.live and the Butte County dashboard. That's
due to them tracking two different things. I've been tracking the state
data, which covers the total hospitalizations for residents and
non-residentts, and Butte County has been tracking residents only.

Both are useful to track, so now the dashboard includes a counter and a
graph covering hospitalized residents.