Rework some text to add some clarity.
The hospitalizations headers are now a bit more clear. We have "Total
Hospitalized", "Total In ICU", and "Butte County Patients". The latter
is hopefully easier to understand than the text I had before.

The age range timeline graph also has a header stating that the more
data was provided on June 9th.