Add clarification to the Skilled Nursing Facilities data.
SNF data reported to (and made availabel by) the state doesn't include
precise case counts. If the value is under 10 cases, they're allowed to
report 1, meaning that we only show 1 case even if there's 9 cases at a
location.

Since this is non-obvious, it's worth adding some clarification. There's
now a blurb about this above the graph, and the tooltips for the graph
data include an "or more" after the count.