Show 7-Day Test Positivity Rates instead of an all-time rate.
The 7-Day Test Positivity Rate is more accurate, in that it shows how
many cases have recently come back positive, better mirroring the
current viral situation rather than being weighed down by tests from a
less infected time.

This is the metric used by the state when calculating the spread of the
virus and determining whether the county needs to be placed on a
monitoring list. As such, it's useful to track.

Note that the state may be using testing/case date information we don't
have to determine their own numbers, so this is a best-effort view into
what the state's idea of the positivity rate would be.