Fix up rendering and date range logic to avoid various issues.
Date range updates were applied by activating zoom mode and reaching
into the internals. This was combined with some code to force some
widths to be 100%.

Most of this janky code could be avoided by just setting the min/max X
values for each graph, and setting an explicit class name for the SVG
backing the graph, which we now do. This lets us get rid of all custom
resizing logic, custom zoom logic, and to speed things up all around.

Sadly, we do still have to reach into the internals just a bit in order
to fix rendering of registered grid lines on a graph, but that's still
far less intrusive than what was there before.