Set the graphs to show the last 90 days by default.
As the year drags on, and it's becoming clear more people are viewing
the dashboard on mobile, it's important not to crowd out any new data
coming in on the graphs. The dashboard was showing the entirety of the
timeline, but the old historical information isn't really that useful to
see anymore.

The graphs are now set to show the last 90 days by default. Users can
always change the range shown if they want to see more or less.