Skip Tableau-based datasets if the scraped date is in the future.
On a couple of occasions, the date pulled out from Tableau dashboards
was in the future, breaking data feeds. This may be due to a bad,
unreliable assumption on what index represents a date, or it may be an
error on the part of the people maintaining the dashboards. Either way,
it's avoidable.

The script now checks if the date is in the future, and if it is, it
skips it.