Show more relative numbers for the In Isolation count.
As this continues to grow, the In Isolation count is worth tracking more
clearly. This is the closest thing we have to an Active Cases number.
Instead of just showing a relative value from the previous day, it's now
showing the same relative values shown for Total Cases: Previous day,
8 days, 14 days, and 30 days.