Attempt to show a 14-day case rate and considerations by state.
This adds in a 14-day case rate graph, with a rough indicator of the
value used by the state to consider whether the county belongs on a
monitoring list.

Both this and the 7-day testing graph now show a line pointing to the
rough value being considered by the state. This is going to be different
than reality a bit, since the state has access to dates and other
information (and there's a discrepancy between county dashboard numbers
and their numbers) that it uses to make these considerations.

Because of this, the text on the line and the text above the graph are
very clear that the state uses slightly different information and that
this is a very rough indicator of what they're seeing.