Auto-calculate Y step sizes and clean up the Y axis and gridlines.
Maintaining the step sizes for the graphs was fine when cases were low,
but it's not worth continuing to manually manage going forward. Now,
they're automatically calculated based on the number of desired ticks
and the maximum value shown in the graph, based on different ranges.

The Y axis in general has been cleaned up on each graph. There's no
longer padding at the top, and there's always a tick value that's at
least as large as the maximum value shown in the graph.

The two monitoring-related graphs with custom lines have their gridlines
re-enabled.