Improve the display on smaller (iPhone 5-era) mobile screens.
The counters now better fit on the screen without awkward wrapping. This
is done by decreasing the size of the counter values and reducing the
margins between them.

The "in the" part of the relative values for counters have been replaced
with just a bit extra space. So now it's "<count>  last 7 days". We'll
see how this ends up feeling over the week, but it reduces noise and
makes it easier to fit on smaller screens.