Fix bad graph values coming in for empty test result days.
Due to a bad calculation, we were getting raw graph data of a - and a +,
resulting in a value of 0. Fortunately, this means nothing was shown in
the graph, but the tooltip showed the broken numbers. The bad
calculation was due to values of 0 on those days, causing relative
values to be computed. We now ignore these days.