Add the California SNF dashboard to the sources list.
Some people may be less interested in CSV files than I am. This links to
the SNF dashboard page, so people can get access to the data and
information directly (as much of a mess as that interface is).

It also adds a "(Dashboard)" suffix to some other link names.