Generate the timeline CSV/JSON from the main Timeline Data sheet.
The way things used to work was that all data was consolidated on a
Timeline Data sheet in Google Sheets, and from that data was normalized
in a pivot table in an Export Data sheet, which produced a CSV file that
`build-datasets.py` would pick up and normalize into a resulting CSV and
JSON file.

This involved a lot of steps, additional processing time in Google
Sheets any time anything updated, and was really painful to manage
whenever new data had to be added to the generated files (especially
given Google Sheet's buggy custom column mode for pivot tables). When
that was originally put into place, `build-datasets.py` was little more
than a fancy downloader and converter.

Now, the Timeline Data is used for export. It contains the CSV column
tags that were previously in the Export Data sheet. `build-datasets.py`
now knows how to skip the extra header rows in that sheet, and how to
handle delta, integer, and percentage values, helping centralize the
work that previously had to be done in the pivot table's custom column
formulas.

This is far more maintainable and will let me quickly update things
going forward whenever new information is added.