Add graphs and counters for the county jail data.
This adds a new section on the county jail. There are counters for the
inmate population, total inmate tests, current inmate cases, percentage
of inmates who are positive, the total staff tests, and the total staff
cases.

Timeline graphs are available showing the inmate popuation and current
cases, and the staff total cases.