Add fine-grained data and date management, handling null data values.
Recently, with the CalREDIE glitch, the county has had a major backlog
to deal with, and that's meant that only some bits of data have been
released this week. That's skewed values terribly, creating obscene test
positivity rates and really messing with statistics and graphs.

This stems from treating all county data as current, meaning that I've
had to include stale values in each day in order to continue updating
for the number of confirmed cases.

Previously, the dashboard assumed there were only a few real groups of
data:

* County-provided data (cases, deaths, isolation, tests, hospitalized
  residents, age groups, regions, and a calculated test positivity rate)
* State-provided hospital data (per-hospital counts, total currently
  hospitalized, and currently in ICU)
* Jail-provided data

Now, it's broken down into:

* County-provided confirmed cases (total and new)
* County-provided deaths (total and new)
* County-provided age groups
* County-provided regions
* County-provided resident hospitalizations
* County-provided testing (total and new results)
* County-provided isolation numbers (total, released)
* State-provided hospital data (total admitted and in ICU)
* State-provided per-hospital data
* Jail-provided data
* Calculated test positivity rate

Only the data within each group needs to be provided. The latest
provided rows for each group are stored, and dates are calculated from
those (which are now represented as a `bc19-o-update-day` CSS class with
a corresponding `data-key="..."` mapping to the key for the latest data
row the count should be pulled from).

(There's also some cleanup in here to make all this more manageable.)

The end result is that I can leave out data the county can't provide
without breaking the dashboard or filling in stale values. Graphs and
counters will all reflect this. Test positivity rates and other
calculations won't be completely busted as a result.