Add a data scraper for the Butte County Jail.
This format has been somewhat stable for a while, but it's unclear how
fluid it really is. Still, this parser scrapes the useful information
from the Butte County Jail page
(https://www.buttecounty.net/sheriffcoroner/Covid-19), providing JSON
and CSV datasets that can be pulled in for analysis and tracking.