Add safer, race-free writing of datasets.
Depending on when the datasets were being generated and when they were
being pulled in (by the website or spreadsheet), the file could end up
blank or corrupt, since the file would have been cleared away and in the
process of writing.

We now safely write these by writing to a file with a temporary
filename, and then moving that over to the destination file, ensuring
that it's only replaced once there's finalized content.