Support looking up Tableau data when an aliasIndices is a negative number.
In Tableau, a column may provide an aliasIndices value that is a
negative number. This has always been a mystery to me, but it appears
that it's actually a reference to a string representation of a value,
bypassing the standard data dictionary for its type. A negative index
appears to be a a negated 1-based index into the cstring data
dictionary.

This change adds support for these, and a comment explaining it.