Add state monitoring tier information to the timeline datasets.
California introduced a new state monitoring tier plan last week. Over
the weekend, I built a parser for this so it can be tracked long-term,
which is now tracked in the master data tracker. This change pulls that
information into the timeline exports, which will be used on the
dashboard.