Update to the SNF parser to handle the recent changes to the dataset.
The upstream dataset redid their tracking a bit, which seems to have
added in some missing numbers, slug identifiers for the facilities, and
a few extra dates at the beginning of the timeline. This change updates
to skip those initial days (which are 0 cases anyway, and would break
any date alignment).

We don't make use of the new information, and don't need to bump the
version of the exported dataset.