Update the county dashboard By Age parsing to get total cases by age.
The county dashboard has intended for their By Age graph to show both:

1) Cases by age *without* deaths
2) Deaths by age

It historically had included deaths in the first bar, but not
consistently. This officially changed, and they updated their graph
format to help make it more clear.

This change works around this to continue getting a total count that
includes cases and deaths for our tracking.