Begin parsing and storing state cases by age.
These counts are always going to be a bit different from the county's,
due to differences in when data is recorded and due to backdating of
cases.

The state's counts are used for county tiers, and to better help
anticipate and validate those counts, I need to pull in the state cases
and start tracking them. These won't be recorded on the dashboard, but
will be available for my own information.