Add a parser for the Chico Unified School District cases spreadsheet.
This parses the CUSD spreadsheet's CSV export, turning it into a
structured JSON document and a CSV file that can be used for analysis.

It's unclear how often the spreadsheet is updated (or whether it'll stay
permanently available -- the webpage embedding it seems to be down). I'm
pulling down daily, and stamping it with the current date, until I know
otherwise.