Update the county dashboard parser for the new age ranges.
The county split up the 0-17 age group into 0-4, 5-12, and 13-17. This
is a great change, but unexpected, and broke parsing.

A lot of work needs to be done to put this information in the dashboard,
and it won't happen today, particularly with the Google Sheets
breakages. So for now, I'm recording that data in the JSON file, but
also consolidating the numbers back into a 0-17 for the dashboard.