Replace the per-region ICU availability export with a Stay-at-Home export.
This new dataset is more comprehensive than the old one, built from the
state data found at this dashboard:

https://public.tableau.com/views/COVID-19Planforreducingcovid-19wregionsmap/regionalmap

The new dataset contains per-region breakdowns with:

* Stay-at-Home active state
* Stay-at-Home effective date
* ICU bed availability percentages
* Adjusted/unadjusted Case rates
* Test positivity rates

Plus statistics on:

* Populations in each monitoring tier
* Number of counties in each Stay-at-Home state
* Number of regions in each Stay-at-Home state
* Population (number and percent) in each Stay-at-Home state

Most of this is for posterity. It's easy to collect now, so why not.

This is still used to build the `state-region-icu-pct.csv` export.