Update the Butte County dashboard parser for probable case/death changes.
Today, the county updated the dashboard to include a graph on probable
cases, and to split actual/probable covid deaths into their own graph.
This last graph is currently missing some age groups (likely due to them
being in the 0-17 ranges and not having any current deaths), so some
work was needed to handle missing age keys.

We now store probable cases and deaths in the JSON file. This isn't
currently in the CSV export.