Store/report on the reported cases/death totals separately from graphs.
Historically, the counters and graphs for cases and deaths used the same
data. This meant that I had to wait a day before backdating any case or
death information in order to convey the increase since the last report.
This affected graphs and was a bit misleading.

Now, those report totals are stored separately, and are used for the
counters. They aren't backdated. This will mean that the deltas aren't
always going to reflect reality exactly (particularly if a number is
bumping up against a weekend), though I'll probably just need to address
that by removing those values periodically from the spreadsheet.

If a value isn't present, the graph data is used.

Along with this, the deaths are now shown relative to 1 day, 7 days, 14
days, and 30 days ago.