Add a flag for toggling Butte County dashboard date behavior.
This week, the dashboard has flip-flopped a couple times between showing
the current day and the previous day in the report. Rather than
reworking the logic each time, there's now just a flag I can toggle in
the code for setting the new behavior. Date calculations will be made
accordingly.