Parse the new vaccine information from the Butte County dashboard.
The county added information on vaccines to the dashboard. This
includes:

* Total allocation
* Total administered
* Total received
* Total first doses ordered
* Total second doses ordered

This change parses this and stores it along with the rest of the
dashboard data.

This information has its own timestamp, and may not be updated daily. As
a result, we're still keying off the main data from the dashboard report
timestamp, and including an "as of" timestamp for the vaccines in that
report's information.