Add a new dataset containing precompiled data for the dashboard.
This is step 1 in improving the foundation of the dashboard. The
original design loaded in all timeline data from the spreadsheet,
compiling the data, and then feeding that into counters and graphs.

There's an overview of that approach here:

https://twitter.com/chipx86/status/1370353813278748677

This change works toward the new design by adding a new
`bc19-dashboard.json` dataset that contains the compiled data. This is
generated server-side, and can be updated to include data from other
datasets.

As of this commit, the dashboard still uses `timeline.json`. That will
be updated separately, so there's a transition period, allowing the
server to generate the new data without breaking the website.