Simplify the appearance of the site and add navigation.
This adds a long-overdue tweak to the website. It removes the gutter
around the main page content, giving more room to the content, and it
tones down the header color.

The main page header's information is now simplified, focusing on
navigation links and not sentences with inline links.

The risk level has been moved to the very top of the page.