Update the look and feel of the various timeline graphs.
When the default date range of the graphs were expanded, and as new data
points were added over time, the bar graphs ended up becoming more
difficult to really look at. This was worse on mobile.

This is the first attempt at really improving this. Most graphs have
been converted to area-step graphs, giving them a clean appearance. The
"new case/death" graphs are still bar graphs, since that better conveys
the addition of new data points.

The "% of Cases In Each Batch of Test Results" graph has been given
wider bar graphs and new colors, to better show the representation of
new cases.

The "Test Results" graph now shows the tests as a step area, and the
results as a bar graph. This sort of creates a nice outline-type effect
that better compares and contrasts the two.