Fix a calculation with the current monitoring tier.
The data feed I'm pulling from provides updates only when the tiers
change, ordered from most recent date to least. The way I was parsing
this, and attempting to fill in the blanks, ended up with the wrong tier
being grabbed from the dataset.

The new approach is to build up a full history of dates-to-tiers, in the
order listed in the dataset, filling in any gaps. This properly
normalizes that data for inclusion in the exported dataset.