Update the dashboard to show the new vaccination information.
The old vaccination counters have been removed, replaced with new % of
population counters for 1+ dose and full doses.

A graph has been added to track the number of 1+ dose and full doses
administered. This is slightly different from the % of County graph,
though not much different. However, it does scale differently, helping
to visualize the rates of vaccination.

Another graph better showing those rates appears below it, showing the
number of new vaccinations over the prior 7 days for each date on the
timeline.