Clean up some of the labels around vaccine doses.
I'm attempting to determine the best way to really convey the difference
between the 1+ doses and the full doses distinctions in the county data.
For the moment, I'm changing these to more explicitly say "Received 1 or
More Doses" and "Fully Vaccinated".

I might end up changing the data to separate this out into a clear
"Partially Vaccinated" (full - 1+ dose) and "Fully Vaccinated" in a
future change.