Update the Butte County dashboard parser for the new once-a-week changes.
Butte County has moved to a once-a-week model, and as part of this,
they've dropped the datestamps on the page saying when parts are
updated and dropped the testing information (delegating to the state
page instead).

We don't have testing information anymore, but the state has that data.
I'll have to pull that out separately.