Switch the dashboard to using case totals and not per-report totals.
At one point, to give a better idea of new cases added over the weekend,
the dashboard was updated to show relative case counts based on report
date and not based on the case date. This aligned better with the
county's way of tracking cases (report date and not incident date).

Now that the county has changed their mechanism, and I've changed my
reports, it makes sense to change the dashboard to also reflect incident
dates. It's now showing "Total Cases" based on `confirmed_cases:total`
and not `confirmed_cases:total_as_of_report`.