Change the default time range to 280 days, and improve date logic.
The timeline graphs now show 280 days worth of data, providing better
context around the current surge.

The logic for computing time ranges is now a bit more standardized.
There's room for improvement, but it should reduce maintenance going
forward.

One big win is that the "selected" value for the range selector in the
DOM is now the source of truth for the default value.