Add high-level overview graphs of cases in schools.
This adds graphs to the dashboard that shows cases in schools during the
school year, both an area chart for totals for the school year and a bar
chart for new cases.

This section appears right after the main dashboard section, since we'll
want to draw attention to it. It will be followed up with counters, and
eventually links to a schools-specific dashboard.