Increase graph sizes and add new tick step size ranges.
The Medium and Standard graphs are now identical in size and tick
counts, helping to consolidate the look of the page. Small and Very
Small, which are also the same, are now taller, helping some data show
through better.

Tick step size calculations have been updated to account for larger data
ranges on the page.