Update to billboard.js 3.1.5.
This is a pretty major upgrade for the dashboard. This release fixes
some issues I previously had to work around, updates to a newer version
of D3.js, and improves performance all throughout.

As part of this, some workarounds and options have been removed.

Data points are now shown when hovering over the graph, now that they
can be set to only show up on hover.

Data is now loaded in through `.load()` instead of `.flow()`, which
seems to be more efficient.

And last, but not least, we're triggering renders of the next graph
after `requestAnimationFrame()` instead of `setTimeout()`.