Turn off stack normalization for vaccination charts.
Some of the vaccination charts were dealing in percentages, and had
stack normalization on in order to ensure the range covered 0-100%.

This was actually very very slow, contributing to slow render times.
Since we're already dealing with percentages to begin with, we don't
need to normalize. (Though, if data is missing, it will stick out on the
graph.)

This turns off normalization, reducing full dashboard load times by
several seconds.