Don't fail if there isn't enough data to satisfy relative counter values.
Counter setup code was asserting if there were more relative counter
values to list than there was data. While this was a reasonable
assumption with months of dashboard data, we don't have that luxury with
some newer datasets.

We now assert that we don't have too few for the data being provided.