Add a dataset builder for the schools dashboard data.
This generates all the data needed to display the schools dashboard.
Since that dashboard has multiple modes (county-wide, per-district, and
per-school), data has to be generated for each level. Further, these are
generated per-school year.

This requires a lot of data to be collected and carefully organized
up-front before generating JSON data from it. It's quite a different
pattern from the main dashboard dataset builder.

As the resulting file size is still pretty big, the next step will be to
generate additional datasets per school year.