Add a notice on the school districts list about unlisted districts.
A common question that I've had so far, and expect to have a lot going
forward, is why some school or district is not listed.

The header for the list of districts now attempts to address this by
specifying that any districts not listed are likely not providing public
data.

There's also a static notice docked at the top of the page that mentions
this, and also tells districts how to contact me if they want to work
with me.