Add Butte College and Nord Country School to the Schools dashboard.
Both schools are now reporting data.

Butte College, unfortunately, is reporting it in a way that loses
historical information and will be hard to track going forward, but it's
something. They're showing just a 7-day and 14-day case count, updated
presumably weekly.

Nord Country School reports by grade level, which is currently not
represented in the generated data, but is tracked for future updates.