Combine, don't replace, per-school case counts for the same day.
For some school "districts" (Nord Country School in particular), a
single school may appear more than once (such as when multiple grades
show cases). This hasn't really come up much before, but is now due to
the Omicron spike.

When this happened, the latest entry would win, replacing any counts
from the previous.

Now, the sum of those entries for the day are used in the data.