Switch the skilled nursing facilities feed to the state feed.
Historically, I've been using the LA Times's data feed for skilled
nursing facilities. This sometimes isn't updated for a few days, making
tracking difficult.

The state now provides a feed. It's a bit different in structure, but
it isn't too hard to get it into the right form for our dataset.

This change switches to that feed.