Update the county dashboard scraper for today's major changes.
The county announced today that they're moving to a twice-a-week
reporting model, and are scrapping some data. Currently, the isolation
and region information is gone, with potentially more to follow. Along
with this, their dashboard has been rebuilt, with entity IDs all
changing.

This change updates the entity IDs, disables scraping of the region and
isolation information, and makes many of the other fields conditional in
order to avoid further breakages.

It also changes the deaths chart scraping to generate its own totals,
since those have been removed.