Remove stale dashboards and update for bi-weekly updates.
The county is moving to bi-weekly updates, Tuesdays and Thursdays. They
also removed the per-region information and isolation numbers.

To set expectations, there's now a banner at the top of the page setting
the update frequency for some data and mentioning some of the
information that's no longer there.

Graphs that are updated bi-weekly now include subtitles with that update
frequency. The By Region graph now states the data is no longer being
updated.

Some older graphs have been removed. Notably, the vaccine demographics
and the adult care facilities.