Update the dashboard for wastewater management.
We now pull down CHHS wastewater data on Chico and Oroville, perform
CDC-described WVAL (Wastewater Viral Activity Level) calculations, and
graph it on the dashboard.

The WVAL calculations involve calculating a baseline from the previous
year in 6 month incremenets, computing the number of standard deviations
from each log-transformed wastewater concentration value in the dataset
deviating from the baseline, converting back to a linear scale, and
using that as a number for WVAL. Those then get mapped to level names
and colors.

We perform these calculations, put the data back into the dashboard
data, and then display it as special text-only counters showing the WVAL
severity name and the timeline graph.

Since we're not updating timeline data anymore from the spreadsheet,
we're no longer using that as the source of truth for the dates. Now,
the date graph data is extended based on the last-updated wastewater
date.