Update for dark mode.
The dashboard is now shown in dark mode, using a dark-blue aesthetic.
This helps the wastewater level colors pop. And it just makes the whole
site a bit nicer to look at.

As part of this, a lot of the prior UI styling has been removed, keeping
this really simple and just about the data.